Add future scopes, clarify more discussion
Fix a reference in read:image to possibly using scopes by data
release, since we're definitely not doing that.  Fix the ordering
of the scopes.  Add expected future scopes and some notes about
how we may change some of the scopes going forward.

Go back to using the ref tag without overriding the title, since
we're not using section numbers (which is what made that look
weird).